Add option for generating an executable that uses the C ABI
Using the `--c-abi` flag causes the compiler to generate an executable
that uses the C ABI for function calls instead of LLVM's Fast Function
Call convention. Using the C ABI means that C functions can call Pony
functions in a Pony program as long as they use the correct mangled
version of the function name.

The C ABI is potentially slower than the Fast Function Call
convention, so there may be a performance impact for executables that
are generated with this option.