Add `_final()` method to deserialized objects that should have one
Objects that were created via deserialization did not have their
finalizer set even if they were of a class that has a `_final()`
method. Consequently, any cleanup code in the `_final()` method would
not be run when the object was garbage collected.

This change checks the type for a finalizer during derialization and
sets the object's `final` pointer if the finalizer exists.