Add C FFI call that allows the setting of an actor's batch size
Within Wallaroo, some actors are going to get much more work
than others. We'd like to be able to allow them to change their
batch size in order to do more work before giving up the scheduler.

This modification adds a c-ffi call that allows an actor to do that.
If an actor wanted to change its batch size to 1000, it could call:

@pony_batch_size[None](this, USize(1000))