Replace "noisey" actor flag with "special"
We haven't used noisey for a while. Swap it out with the "special"
flag. This can be used from Pony code to mark an actor as special
and we can then check for that flag when getting custom telemetry
data from Ponyc.

Not useful without doing ad-hoc metrics collection in ponyc.

Will remove eventually.