Make chunk_t fit into 32 bytes on 64 bit platforms
This commit changes how small and large chunks are implemented
in order to fit the chunk structs into 32 bytes on 64 bit platforms.
This is accomplished by encoding information into the high 12 bits
of `m`, `actor`, and `next` as required.

This required splitting chunk_t into small_chunk_t and large_chunk_t
each with a different definition on 64 bit platforms. chunk_t was
also split on 32 bit platforms but the only difference between the
two there is the pointer type for the `next` field.

Additionally, a lot of actions that used to be directly done on
fields of chunk_t are now hidden behind getters/setters in order
to use the same code on both 32 and 64 bit platforms. The getters
and setters have `ifdefs` as required.