Fix file package issues on OSX and enable large files on 32 bit
Prior to this commit the file package used `lseek64` which doesn't
exist on OSX/bsd. This commit fixes linux to enable the
`_FILE_OFFSET_BITS=64` macro to tell linux to have the standard
posix file related calls behave same as for OSX/bsd and use a
64 bit `off_t` even on 32 bit platforms. This also resulted in
being able to use `lseek` on all platforms with a `I64` return type.